added referral user to publisher signup both in db and in the sent mail

+++ apparelrow/dashboard/middleware.py
@@ -2,22 +2,16 @@ from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.core.signing import BadSignature
 
+from apparelrow.dashboard.utils import get_referral_user_from_cookie
 
 class ReferralMiddleware(object):
     def process_response(self, request, response):
-        if settings.APPAREL_DASHBOARD_REFERRAL_COOKIE_NAME in request.COOKIES and request.user and request.user.is_authenticated():
+        if settings.APPAREL_DASHBOARD_REFERRAL_COOKIE_NAME in request.COOKIES and hasattr(request, 'user') and request.user and request.user.is_authenticated():
             if not request.user.referral_partner_parent:
-                try:
-                    user_id = request.get_signed_cookie(settings.APPAREL_DASHBOARD_REFERRAL_COOKIE_NAME)
-                    try:
-                        user = get_user_model().objects.get(pk=user_id)
-                        if user.is_active and user.is_partner and user.referral_partner:
-                            request.user.referral_partner_parent = user
-                            request.user.save()
-                    except get_user_model().DoesNotExist:
-                        pass
-                except BadSignature:
-                    pass
+                user = get_referral_user_from_cookie(request)
+                if user and user.is_active and user.is_partner and user.referral_partner:
+                    request.user.referral_partner_parent = user
+                    request.user.save()
 
             response.delete_cookie(settings.APPAREL_DASHBOARD_REFERRAL_COOKIE_NAME)
